+++ devices/metal/metal_heap.h
@@ -26,10 +26,10 @@ OIDN_NAMESPACE_BEGIN
     void init();
     void free();
 
+    MetalEngine* engine;
     id<MTLHeap> heap;
     size_t byteSize;
     Storage storage;
-    MetalEngine* engine;
   };
 
 OIDN_NAMESPACE_END